implement session and change database model

+++ eTrener/eTrener/Models/ProductModel.cs
@@ -11,9 +11,9 @@ namespace eTrener.Models
         [Key]
         public int ProductId { get; set; }
         public string Name { get; set; }
-        public double Calories { get; set; }
-        public double Fat { get; set; }
-        public double Carbs { get; set; }
-        public double Protein { get; set; }
+        public decimal Calories { get; set; }
+        public decimal Fat { get; set; }
+        public decimal Carbs { get; set; }
+        public decimal Protein { get; set; }
     }
 }
